requirements.txt
crewai>=0.80.0
crewai-tools>=0.15.0
litellm>=1.40.0
langsmith>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
flask>=3.0.0
